Permissions

+++ app/Http/Controllers/CommentController.php
@@ -24,6 +24,10 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        if (auth()->id() !== $comment->user_id) {
+            abort(403);
+        }
+
         $comment->delete();
 
         return to_route('feature.show', $comment->feature_id);
